Split BookPage into modular components and simplify mode logic
- Extracted layout parts: BookPageHeader, BookReadingDetails, BookPersonalStats, BookPersonalNotes, BookPageActions
- Added shared UI atoms: SectionTitle, Detail, Meta
- Simplified mode handling (view/edit/create) directly inside useEffect without redundant booleans
- Introduced barrel export (index.ts) for cleaner imports
