+++ frontend/src/components/commons/book/index.ts
@@ -1,0 +1,8 @@
+export * from "./BookPageActions";
+export * from "./BookPageHeader";
+export * from "./BookPersonalNotes";
+export * from "./BookPersonalStats";
+export * from "./BookReadingDetails";
+export * from "./Detail";
+export * from "./Meta";
+export * from "./SectionTitle";
